Make sampling flexible with parameters as CLI input args
Now any input argument can be passed to the sampling program as
long as they are all at the end of the Python script command.
Moreover, the GGSW vlaue can be choosen. It can either be a fixed
value [0,1] for the whole sampling phase or be randomly chosen
at each repetition. The GGSW is also added to the result files
at the 5th position.