chore(ecosystem): removes concrete and concrete-boolean after split
The original repository (zama-ai/concrete) was split in two
repositories. The existing one (zama-ai/concrete) which keep
hosting the `concrete` and `concrete-boolean` crates, and a new
one (zama-ai/concrete-core) which hosts the other crates.

This commit is the first commit (zama-ai/concrete-core), which removes
the `concrete` and the `concrete-core` repositories from this one.