feat(concrete_core): add serialization support
This commit adds support for serialization in `concrete-core`. Two
engine traits were added to the `specification` module. This trait was
then implemented in both the `default` and the `fftw` backend, for all
existing entities.

In addition, an early support for serialization versioning was added.
All entities are currently tagged with the `V0` version, which is then
checked at deserialization.