chore(ci): set scheduled benchmarks with results sent to ci bot
A new scheduled workflow is created to run concrete benchmarks
using only CPU backend on a weekly basis. The results are then
parsed before being sent to Slab ci-bot.
After that Slab handleq transformation of these results into
Prometheus metrics in order to be displayed in visualization system
which is currently a Grafana instance.